Initial Commit.
This is the first pass at writing a boring todo app... only less boring.
Inspired by a forum post about rules / accomplishments to be made before progressing in Minecraft, this app is a simple checklist of where you are on that journey.

I originally planned on just hacking something together for the sake of it, but that turned into 'why not make a todo app' every dev has one - and this could be yours... yay!
Anyway, heres the first quick code to get something working. Each additional commit / release will be refactoring and improving this code. Effectively treating it as a living
example of being handed legacy code and working to improve it and build something more maintainable and personally enjoyable to work on.

#Goals for this project
- Build a ToDo app that isn't boring.
- Provide live example of working with a 'legacy' codebase
- Write more stuff in swift, since its a great language
- Something else... I'm sure.